50 Startups
It is a dataset containing the data of 50 startups and these data include R & D expenditures, administration expenses, marketing expenses, the state in which the startup is located and the annual profit of the startup.

Business Scenario: We imagine we are working for a Venture Capital Fund that wonders which startup to invest in? And so to help these investors of the Venture Capital Fund to know in which startup to invest, we will build a linear regression model that will be able to understand the correlation between the independent variables and the profit data which is the dependent variable that we are trying to predict.