Social Network - Regression Logistic
Business scenario: We imagine that we are working for a car manufacturer that has just released its new product. This new product is a luxury car that is absolutely not expensive. And to sell this beautiful car, this company has put an advertisement on social networks. And the users of these social networks by seeing the ad decide whether or not to buy the car. Display advertising on social networks represents a certain cost and therefore the company can optimize its costs and therefore its returns on investment since the users it targets will be the most likely to buy the car through this ranking of given probabilities by our classification model. So that's the CHALLENGE, that's what we need to create with logistic regression.