Salary_prediction
We are going to find a correlation between the numbers of year experience and salaries of workers. We want to predict salaries of other workers who have many years experience.

Business Scenario: We can imagine that this dataset is thus a dataset of a company who the manager need to make predictions on workers's salaries in order to make some anticipations. We are going to help this manager to make good predictions.