Prediction of purchase
Business Scenario: We can imagine that this dataset is thus a dataset of a company which has a file of customers, who has information on these customers and especially who has the information to know if the customers bought or not one of these new products . The company tries to correlate country, age and salary with the customer's decision to buy the product. We will therefore use the independent variables to try to predict whether the customer buys the product or not. In other words, with the independent variables, we will try to predict whether the person will buy a product or not.