Red wine quality
Suppose we are at a party and we want to drink the best wine available. We can ask recommendations to our friends but there is a good chance that their opinion is not objective because they are already a little drunk. Fortunately, we brought pH test strips and other tools to measure various physicochemical properties. But as we love Data Science, we decided to accept the challenge and use machine learning to predict wine quality based on its physico-chemical attributes.